+++ src/tools/Built-in/7_sub_command/Control/Control.tools.ts
@@ -1,0 +1,59 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
+import * as vscode from 'vscode';
+import { ControlSubCmdList } from './Control.data';
+
+export const { snippetControl, ControlMDMap } = (() => {
+    const MDMapRW = new Map<string, vscode.MarkdownString>();
+    const snippetListRW: vscode.CompletionItem[] = [];
+
+    for (const v of ControlSubCmdList) {
+        const {
+            SubCommand,
+            body,
+            doc,
+            link,
+            exp,
+        } = v;
+        const upName: string = SubCommand.toUpperCase();
+        const md: vscode.MarkdownString = new vscode.MarkdownString('', true)
+            .appendMarkdown(`Control, ${SubCommand}`)
+            .appendCodeblock(body, 'ahk')
+            .appendMarkdown(`[(Read Doc)](${link})\n\n`)
+            .appendMarkdown(doc)
+            .appendMarkdown('\n\n***')
+            .appendMarkdown('\n\n*exp:*')
+            .appendCodeblock(exp.join('\n'), 'ahk');
+
+        md.supportHtml = true;
+        MDMapRW.set(upName, md);
+
+        const item: vscode.CompletionItem = new vscode.CompletionItem({
+            label: `Control, ${SubCommand}`, // Left
+            description: 'sub-command', // Right
+        });
+        item.kind = vscode.CompletionItemKind.Keyword;
+        const body2: string = body.includes('[')
+            ? body
+                .replaceAll(/\s*,\s*\[/gu, '[')
+                .replaceAll(/\[\s*,\s*/gu, '[')
+                .replaceAll(/\s*\[\s*/gu, ',[$0 ')
+            : body;
+        item.insertText = new vscode.SnippetString(body2);
+        item.detail = 'Control';
+        item.documentation = md;
+
+        snippetListRW.push(item);
+    }
+
+    /**
+     * after initialization clear
+     */
+    ControlSubCmdList.length = 0;
+
+    // ---
+    // ---
+    return {
+        snippetControl: snippetListRW as readonly vscode.CompletionItem[],
+        ControlMDMap: MDMapRW as ReadonlyMap<string, vscode.MarkdownString>,
+    };
+})();